requirements.txt
selenium==4.23.1
requests==2.32.3
pytz==2024.1
python-dateutil==2.9.0.post0
gspread==6.1.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
pandas==2.2.2
lxml==5.2.2
